+++ desietc/accum.py
@@ -307,7 +307,7 @@ class Accumulator(object):
 
         return True
 
-    def get_efftime(self, realtime, signal, background):
+    def get_efftime(self, realtime, signal, background, scale=(0.56 / 0.435)**2):
         """Calculate the effective exposure time corresponding to the specified real
         exposure time, accumulated signal and background rates, and their nominal values
         and MW transparency specified in the last call to :meth:`setup_exposure`.
@@ -315,4 +315,4 @@ class Accumulator(object):
         sig_factor = self.MW_transp * signal
         rdnoise = self.rdnoise_1ks * 1000 / np.maximum(0.1, realtime)
         bg_factor = (background + rdnoise) / (1 + self.rdnoise_1ks)
-        return realtime * sig_factor ** 2 / bg_factor
+        return scale * realtime * sig_factor ** 2 / bg_factor